Flytt hjemler-kodeverk til klage-API under /api/klage/kodeverk/hjemler
